Removing stuff from parsing and moving to TestSuite
